requirements.txt
Flask
python-dotenv
idcube-hive-connector
idcube-python-utils
infrabot-diy-python-client